Improve & track performance, improve data presentation
Add functions to report on how long data acquisition takes. Based on the reports from those functions, target poorly performing areas of data acquisition and improve them. This commit leaves the poorly performing code so that the end-user can test for themselves, based on the setting of a script-level variable $opt.

Testing against O365 is of course dependent on the time of day and many other factors, but in my testing, $opt = $true leads to 60%-70% reduction in wall-clock time for data acquisition.

Also, improve data presentation in several areas (e.g., "5,120 GB" vs "5120.00 GB"). This needs more work, but this commit is a start.

Add "Set-StrictMode -Version Latest" at the top of the script and correct the errors found based on that.